requirements.txt
openai>=1.0.0
python-docx>=0.8.11
pyyaml>=6.0
pymupdf>=1.23.0
pillow>=9.0.0
pytesseract>=0.3.10
